refactor: refactored search terms
Bench: 4959499
